requirementsOthers.txt
tensorflow>=1.7,<1.8
Pillow>=4.2.1
matplotlib
numpy>=1.13.3,<=1.14.5
jupyter
pytest>=3.2.2,<4.0.0
docopt
pyyaml
jsonpickle
protobuf>=3.6,<3.7
grpcio>=1.11.0,<1.12.0
